refining Users

+++ app/Views/users.php
@@ -1,26 +1,37 @@
 <?= $this->extend('layout') ?>
 
 <?= $this->section('extraScript') ?>
-<script src="js/ajax.googleapis.com_ajax_libs_jquery_3.6.4_jquery.min.js"></script>
-<script src="js/cdn.datatables.net_1.13.4_js_jquery.dataTables.min.js"></script>
+<script src="js/jquery-3.1.1.js"></script>
 <?= $this->endSection() ?>
 
 <?= $this->section('main') ?>
 
 <!-- Page Heading -->
-<div class="tm-card-header uk-light uk-margin-remove-left">
-    <div uk-grid class="uk-flex-middle">
-        <div class="uk-width-1-2@m">
-            <h3 class="tm-h3"><?= lang('Global.usersList') ?></h3>
-        </div>
+<?php if ($ismobile === false) { ?>
+    <div class="tm-card-header uk-margin-remove-left">
+        <div uk-grid class="uk-flex-middle">
+            <div class="uk-width-1-2@m">
+                <h3 class="tm-h3"><?= lang('Global.usersList') ?></h3>
+            </div>
 
-        <!-- Button Trigger Modal Add -->
-        <div class="uk-width-1-2@m uk-text-right@m">
+            <!-- Button Trigger Modal Add -->
+            <div class="uk-width-1-2@m uk-text-right@m">
+                <button type="button" class="uk-button uk-button-primary uk-preserve-color" uk-toggle="target: #tambahdata"><?= lang('Global.Adduser') ?></button>
+            </div>
+            <!-- End Of Button Trigger Modal Add -->
+        </div>
+    </div>
+<?php } else { ?>
+    <h3 class="tm-h3 uk-text-center"><?= lang('Global.usersList') ?></h3>
+    <div class="uk-child-width-auto uk-flex-center" uk-grid>
+        <div>
+            <button type="button" class="uk-button uk-button-secondary uk-preserve-color" uk-toggle="target: #filter">Filter <span uk-icon="chevron-down"></span></button>
+        </div>
+        <div>
             <button type="button" class="uk-button uk-button-primary uk-preserve-color" uk-toggle="target: #tambahdata"><?= lang('Global.Adduser') ?></button>
         </div>
-        <!-- End Of Button Trigger Modal Add -->
     </div>
-</div>
+<?php } ?>
 <!-- End Of Page Heading -->
 
 <?= view('Views/Auth/_message_block') ?>
@@ -117,6 +128,115 @@
 <!-- End Of Modal Add -->
 
 <!-- Table Of Content -->
+<?php if ($ismobile === false) { ?>
+    <form class="uk-margin" id="searchform" action="users" method="GET">
+        <div class="uk-child-width-auto uk-flex-between uk-flex-middle" uk-grid>
+            <div>
+                <div class="uk-child-width-auto uk-grid-small uk-flex-middle" uk-grid>
+                    <div><?= lang('Global.search') ?>:</div>
+                    <div><input class="uk-input uk-form-width-medium" id="search" name="search" <?= (isset($input['search']) ? 'value="'.$input['search'].'"' : '') ?> /></div>
+                    <div>
+                        <select class="uk-select uk-form-width-medium" id="rolesearch" name="rolesearch">
+                            <option value="0"><?= lang('Global.allAccess') ?></option>
+                            <?php
+                            foreach ($roles as $role) {
+                                if (isset($input['rolesearch']) && ($input['rolesearch'] === $role->id)) {
+                                    $selected = 'selected';
+                                } else {
+                                    $selected = '';
+                                }
+                                if ($authorize->inGroup('admin', $uid) === true) {
+                                    $position = array('owner', 'superuser', 'guests');
+                                    if ((!in_array($role->name, $position))) {
+                                        echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                                    }
+                                } elseif ($authorize->inGroup('owner', $uid) === true) {
+                                    $position = array('superuser', 'guests');
+                                    if ((!in_array($role->name, $position))) {
+                                        echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                                    }
+                                } elseif ($authorize->inGroup('superuser', $uid) === true) {
+                                    echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                                }
+                            }
+                            ?>
+                        </select>
+                    </div>
+                </div>
+            </div>
+            <div>
+                <div class="uk-child-width-auto uk-grid-small uk-flex-middle" uk-grid>
+                    <div><?= lang('Global.display') ?></div>
+                    <div>
+                        <select class="uk-select uk-form-width-xsmall" id="perpage" name="perpage">
+                            <option value="10" <?= (isset($input['perpage']) && ($input['perpage'] === '10') ? 'selected' : '') ?>>10</option>
+                            <option value="25" <?= (isset($input['perpage']) && ($input['perpage'] === '25') ? 'selected' : '') ?>>25</option>
+                            <option value="50" <?= (isset($input['perpage']) && ($input['perpage'] === '50') ? 'selected' : '') ?>>50</option>
+                            <option value="100" <?= (isset($input['perpage']) && ($input['perpage'] === '100') ? 'selected' : '') ?>>100</option>
+                        </select>
+                    </div>
+                    <div><?= lang('Global.perPage') ?></div>
+                </div>
+            </div>
+        </div>
+    </form>
+<?php } else { ?>
+    <div id="filter" class="uk-margin" hidden>
+        <form id="searchform" action="users" method="GET">
+            <div class="uk-margin-small uk-flex uk-flex-center">
+                <input class="uk-input uk-form-width-medium" id="search" name="search" placeholder="<?= lang('Global.search') ?>" <?= (isset($input['search']) ? 'value="'.$input['search'].'"' : '') ?> />
+            </div>
+            <div class="uk-margin-small uk-flex uk-flex-center">
+                <select class="uk-select uk-form-width-medium" id="rolesearch" name="rolesearch">
+                    <option value="0"><?= lang('Global.allAccess') ?></option>
+                    <?php
+                    foreach ($roles as $role) {
+                        if (isset($input['rolesearch']) && ($input['rolesearch'] === $role->id)) {
+                            $selected = 'selected';
+                        } else {
+                            $selected = '';
+                        }
+                        if ($authorize->inGroup('admin', $uid) === true) {
+                            $position = array('owner', 'superuser', 'guests');
+                            if ((!in_array($role->name, $position))) {
+                                echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                            }
+                        } elseif ($authorize->inGroup('owner', $uid) === true) {
+                            $position = array('superuser', 'guests');
+                            if ((!in_array($role->name, $position))) {
+                                echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                            }
+                        } elseif ($authorize->inGroup('superuser', $uid) === true) {
+                            echo '<option value="' . $role->id . '" '.$selected.'>' . $role->name . '</option>';
+                        }
+                    }
+                    ?>
+                </select>
+            </div>
+            <div class="uk-margin uk-child-width-auto uk-grid-small uk-flex-middle uk-flex-center" uk-grid>
+                <div><?= lang('Global.display') ?></div>
+                <div>
+                    <select class="uk-select uk-form-width-xsmall" id="perpage" name="perpage">
+                        <option value="10" <?= (isset($input['perpage']) && ($input['perpage'] === '10') ? 'selected' : '') ?>>10</option>
+                        <option value="25" <?= (isset($input['perpage']) && ($input['perpage'] === '25') ? 'selected' : '') ?>>25</option>
+                        <option value="50" <?= (isset($input['perpage']) && ($input['perpage'] === '50') ? 'selected' : '') ?>>50</option>
+                        <option value="100" <?= (isset($input['perpage']) && ($input['perpage'] === '100') ? 'selected' : '') ?>>100</option>
+                    </select>
+                </div>
+                <div><?= lang('Global.perPage') ?></div>
+            </div>
+        </form>
+    </div>
+<?php } ?>
+<script>
+    document.getElementById('rolesearch').addEventListener("change", submitform);
+    document.getElementById('search').addEventListener("change", submitform);
+    document.getElementById('perpage').addEventListener("change", submitform);
+
+    function submitform() {
+        document.getElementById('searchform').submit();
+    };
+</script>
 <div class="uk-overflow-auto uk-margin">
     <table class="uk-table uk-table-middle uk-table-large uk-table-hover uk-table-divider">
         <thead>
@@ -130,20 +250,37 @@
         </thead>
         <tbody>
             <?php foreach ($users as $user) { ?>
-                <tr>
-                    <td><?= $user->firstname; ?> <?= $user->lastname; ?></td>
-                    <td><?= $user->username; ?></td>
-                    <td><?= $user->email; ?></td>
-                    <td><?= $user->role; ?></td>
-                    <td class="uk-child-width-auto uk-grid-small uk-flex-center" uk-grid>
-                        <div>
-                            <a class="uk-icon-button" uk-icon="pencil" uk-toggle="target: #editdata<?= $user->id ?>"></a>
-                        </div>
-                        <div>
-                            <a uk-icon="trash" class="uk-icon-button-delete" href="users/delete/<?= $user->id ?>" onclick="return confirm('<?= lang('Global.deleteConfirm') ?>')"></a>
-                        </div>
-                    </td>
-                </tr>
+                <?php if (($authorize->inGroup('superuser', $uid) === true) && ($user->role === 'guests')) { ?>
+                    <tr>
+                        <td><?= $user->firstname; ?> <?= $user->lastname; ?></td>
+                        <td><?= $user->username; ?></td>
+                        <td><?= $user->email; ?></td>
+                        <td><?= $user->role; ?></td>
+                        <td class="uk-child-width-auto uk-grid-small uk-flex-center" uk-grid>
+                            <div>
+                                <a class="uk-icon-button" uk-icon="pencil" uk-toggle="target: #editdata<?= $user->id ?>"></a>
+                            </div>
+                            <div>
+                                <a uk-icon="trash" class="uk-icon-button-delete" href="users/delete/<?= $user->id ?>" onclick="return confirm('<?= lang('Global.deleteConfirm') ?>')"></a>
+                            </div>
+                        </td>
+                    </tr>
+                <?php } elseif ($user->role !== 'guests') { ?>
+                    <tr>
+                        <td><?= $user->firstname; ?> <?= $user->lastname; ?></td>
+                        <td><?= $user->username; ?></td>
+                        <td><?= $user->email; ?></td>
+                        <td><?= $user->role; ?></td>
+                        <td class="uk-child-width-auto uk-grid-small uk-flex-center" uk-grid>
+                            <div>
+                                <a class="uk-icon-button" uk-icon="pencil" uk-toggle="target: #editdata<?= $user->id ?>"></a>
+                            </div>
+                            <div>
+                                <a uk-icon="trash" class="uk-icon-button-delete" href="users/delete/<?= $user->id ?>" onclick="return confirm('<?= lang('Global.deleteConfirm') ?>')"></a>
+                            </div>
+                        </td>
+                    </tr>
+                <?php } ?>
             <?php } ?>
         </tbody>
     </table>
@@ -254,12 +391,4 @@
     </div>
 <?php } ?>
 <!-- End Of Modal Edit -->
-
-<!-- Search Engine Script -->
-<script>
-    $(document).ready(function() {
-        $('#example').DataTable();
-    });
-</script>
-<!-- Search Engine Script End -->
 <?= $this->endSection() ?>
